+++ src/classic/avrcp.c
@@ -67,10 +67,10 @@ typedef struct {
 static void avrcp_packet_handler(uint8_t packet_type, uint16_t channel, uint8_t *packet, uint16_t size);
 static void avrcp_start_next_sdp_query(void);
 
-static const char * avrcp_default_controller_service_name = "BTstack AVRCP Controller Service";
-static const char * avrcp_default_controller_service_provider_name = "BTstack AVRCP Controller Service Provider";
-static const char * avrcp_defaul_target_service_name = "BTstack AVRCP Target Service";
-static const char * avrcp_default_target_service_provider_name = "BTstack AVRCP Target Service Provider";
+static const char * avrcp_default_controller_service_name = "AVRCP Controller";
+static const char * avrcp_default_controller_service_provider_name = "BlueKitchen";
+static const char * avrcp_default_target_service_name = "AVRCP Target";
+static const char * avrcp_default_target_service_provider_name = "BlueKitchen";
 
 static const char * avrcp_subunit_type_name[] = {
         "MONITOR", "AUDIO", "PRINTER", "DISC", "TAPE_RECORDER_PLAYER", "TUNER",
@@ -348,7 +348,7 @@ void avrcp_create_sdp_record(bool controller, uint8_t * service, uint32_t servic
         if (controller){
             de_add_data(service, DE_STRING, (uint16_t) strlen(avrcp_default_controller_service_name), (uint8_t *) avrcp_default_controller_service_name);
         } else {
-            de_add_data(service, DE_STRING, (uint16_t) strlen(avrcp_defaul_target_service_name), (uint8_t *) avrcp_defaul_target_service_name);
+            de_add_data(service, DE_STRING, (uint16_t) strlen(avrcp_default_target_service_name), (uint8_t *) avrcp_default_target_service_name);
         }
     }